backend/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
mysqlclient==1.4.6
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==0.16.1
wincertstore==0.2
WTForms==2.2.1
zipp==3.1.0
